requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Jinja2==3.1.6
python-iso639==2025.2.18
openai==1.106.1
pydantic-settings==2.10.1
